requirements.txt
pandas
openpyxl
python-docx
python-docx-replace
